% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_synoptReg.R
\name{as_synoptReg}
\alias{as_synoptReg}
\title{Convert a SpatRaster object to a data.frame structure required by synoptReg}
\usage{
as_synoptReg(x)
}
\arguments{
\item{x}{SpatRaster time series containing the atmospheric data (e.g. mean sea level pressure).}
}
\value{
A data.frame with the following variables: \code{x, y, time, value, var, units}
}
\description{
Convert a SpatRaster from the terra package to a data.frame formatted in the way
required by synoptReg .
}
\examples{
# Load data
slp_file <- system.file("extdata", "mslp_ei.nc", package = "synoptReg")

# Reading data
slp <- read_nc(slp_file)

# Converting it to the required structure by synoptReg
slp_df <- as_synoptReg(slp)

}
\seealso{
\code{\link{read_nc}}
}
