% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncep.R
\name{download_ncep}
\alias{download_ncep}
\title{Download NCEP/NCAR data}
\usage{
download_ncep(var = "air.2m", level = "gaussian", month_range = c(1,
  12), year_range = c(2010, 2017), lat_range = c(30, 60),
  lon_range = c(-30, 10), dailymean = TRUE, hour = NULL,
  reanalysis2 = TRUE, save_download = TRUE, file_name = NULL)
}
\arguments{
\item{var}{slp 'sea level pressure' (default) for more variables see help of ?NCEP.gather}

\item{level}{surface (default)}

\item{month_range}{min,max month c(1,12) (default)}

\item{hour}{One hour of the following: 0,6,12 or 18}
}
\value{
a list with: itemize{
   \item{a matrix with the reanalysis data}
   \item{a date-time object}
   \item{a data.frame with longitude and latitude for each point}
}
}
\description{
Weather Data from NCEP/NCAR Reanalysis via RNCEP package
}
