% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulation_types_calculation.R
\name{\%>\%}
\alias{\%>\%}
\alias{classjen}
\title{Circulation Types Calculation}
\usage{
classjen(points, month_range = c(1, 12), year_range = c(2010, 2017),
  lat_range = c(30, 60), lon_range = c(-30, 10),
  press_dailymean = TRUE, hour = NULL, reanalysis2 = TRUE,
  save_download = TRUE, load_data = FALSE, file_name = NULL, ...)
}
\arguments{
\item{points}{16 coord point pair in order 1 to 16. More in Details.}

\item{month_range}{Numeric. Specifies the range of months from each year.
c(1,12) or 1}

\item{year_range}{Numeric. Specifies the range of years.
c(2010,2017) or 2010}

\item{lat_range}{Numeric. Specifies the range of latitude.
c(30,60)}

\item{lon_range}{Same as lat_range.
c(-30,10)}

\item{press_dailymean}{If you want to the use odaily mean pressue TRUE, or if FALSE you can use a specific hour.}

\item{hour}{One or several hours of the following:
0,6,12 or 18 c(0,6) or 1}

\item{reanalysis2}{default is TRUE. See ?NCEP.gather.}

\item{save_download}{If TRUE, than the downloaded pressure array is saved as RData file.}

\item{load_data}{If you already have downloaded data, format has to be the same as the returned array from the download function.}

\item{file_name}{Name of downloaded data file.}
}
\value{
a data.frame with CT for each requested date and time
}
\description{
Based on Lorenzo et al. (2008): Links between circulation weather types and teleconnection
}
