% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lamb_scheme.R
\name{plot_lamb_scheme}
\alias{plot_lamb_scheme}
\title{Plot Lamb Scheme}
\usage{
plot_lamb_scheme(points)
}
\arguments{
\item{points}{points obtained from the \code{get_lamb_points} function.}
}
\value{
a ggplot map.
}
\description{
Visualize the Lamb Scheme
}
\examples{
points <- get_lamb_points(x = -5, y = 40)

plot_lamb_scheme(points)

}
