% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_pca.R
\name{raster_pca}
\alias{raster_pca}
\title{Raster PCA}
\usage{
raster_pca(x, aggr = NULL, focal = NULL)
}
\arguments{
\item{x}{SpatRaster.}

\item{aggr}{Integer. Aggregation factor based on function \code{aggregate} of \pkg{terra} package.}

\item{focal}{Integer. smooth filter based on function \code{focal} of \pkg{terra} package.}
}
\value{
a list with: \itemize{
   \item{A SpatRaster containing the results of the PCA }
   \item{A data.frame containing the main results of the PCA (standard deviation, proportion of variance and cumulative variance}
}
}
\description{
Perform a Principal Component Analysis on a SpatRaster
}
