% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalization.R
\name{regionalization}
\alias{regionalization}
\title{Environmental regionalization}
\usage{
regionalization(x, centers, iter.max = 100, nstart = 100, algorithm = "Lloyd")
}
\arguments{
\item{x}{SpatRaster}

\item{centers}{Integer. Number of clusters.}

\item{iter.max}{Integer. The maximum number of iterations allowed. Default 100.}

\item{nstart}{Integer. How many random sets should be chosen? Default 100.}

\item{algorithm}{Default Lloyd.}
}
\value{
a list with: \itemize{
   \item{A list with the results of the K-means performance}
   \item{A SpatRaster with the final regionalization }
   \item{A list with the results of the K-means performance}
}
}
\description{
Perform an unspervised clustering of a SpatRaster
}
