% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_lamb_clas.R
\name{spatial_lamb}
\alias{spatial_lamb}
\title{Spatial objective Lamb Weather Type Classification}
\usage{
spatial_lamb(
  msl,
  xmin = -180,
  xmax = 180,
  ymin = -80,
  ymax = 80,
  U = T,
  thr = c(6, 6),
  cores = 1
)
}
\arguments{
\item{msl}{Mean Sea Level pressure gridded data.}

\item{xmin}{minimum longitude}

\item{xmax}{maximum longitude}

\item{ymin}{minimum latitude}

\item{ymax}{maximum longitude}

\item{U}{Logical. If T, Jones et al. 2013 approach is applied, maintaining the U-type in the classification. If F, U is removed as detailed in Trigo and DaCamara, 2000.}

\item{thr}{threshold used for Unclassified days (total shear vorticity and total flow, respectively). Default c(6,6).}

\item{cores}{Number of cores to be used}
}
\value{
A list with: \itemize{
   \item{A data frame containing the gridded  classification}
}
}
\description{
Calculates the  automatic Lamb or Jenkinson and Collison classification at each grid point. 
The approach details are described in  \code{\link{lamb_clas}}.
}
\examples{
library(dplyr)
data(msl)

msl <- filter(msl, time >= "2000-01-01", time <= "2000-03-30")
clas <- spatial_lamb(msl, xmin = 5,xmax = 15, ymin = 40, ymax = 50, cores = 1)

}
\references{
{
Jenkinson, A.F., Collison F.P (1977)
\emph{An initial climatology of gales over the North Sea}
Synoptic Climatology Branch Memorandum, No. 62.Meteorological Office: Bracknell, England.

Jones, P. D., Hulme M., Briffa K. R. (1993)
\emph{A comparison of Lamb circulation types with an objective classification scheme}
Int. J. Climatol. 13: 655–663.

Jones, P. D., Harpham C, Briffa K. R. (2013)
\emph{Lamb weather types derived from Reanalysis products}
Int. J. Climatol. 33: 1129–1139.
}

Trigo, R., DaCamara C. (2000)
\emph{Circulation weather types and their impact on the precipitation regime in Portugal}
Int. J. Climatol. 20: 1559-1581.
}
\seealso{
\code{\link{lamb_clas}}
}
