% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalization.R
\name{regionalization}
\alias{regionalization}
\title{Environmental regionalization}
\usage{
regionalization(raststack, centers, iter.max = 100, nstart = 100)
}
\arguments{
\item{raststack}{Raster Stack.}

\item{centers}{Integer. Number of clusters.}

\item{iter.max}{Integer. The maximum number of iterations allowed. Default 100.}

\item{nstart}{Integer. How many random sets should be chosen? Default 100.}
}
\value{
a list with: \itemize{
   \item{A raster with the final regionalization }
   \item{A list with the results of the K-means performance}
   \item{A raster displaying a pseudo-MAE error based on the difference between each pixel value and its respective centroide}
   \item{A numeric pseudo-MAE mean value for the entire map}
}
}
\description{
Perform an unspervised clustering of the Raster Stack
}
